[LL] stub-based deserializer: use compiled stubs
Compiled stubs are a source of truth and have exhaustive information
which might not be available from decompiled stubs.

This commit also forces the Standalone to use real stubs instead of
fake ones. It is required since stubs are used in the test
infrastructure. They have to be real to test real behavior.
Also, the index can be built lazily from now.

Note that this commit cannot be split on [LL] and [Standalone] part
since both changes depend on each other.

Changes inside `lazyResolveBuiltinsBinary/Int.txt` caused by
inconsistencies due to `getContainingKtFile().isCompiled()` checks
inside psi implementations. For instance, `KtProperty.getInitializer`
returns `null` for decompiler files if the stub is present. If the AST
is already loaded, then it would work as usually via AST search

^KT-79798